+++ src/lib/validators/measurement.ts
@@ -1,0 +1,31 @@
+import { z } from "zod";
+import { ISOStringToSQLTimestamp } from "lib/units/convertTimestamp";
+
+export const zCreateMeasurement = z.object({
+  time: z
+    .string()
+    // maximize compatibility and validate the inputted date
+    .transform((str: string) => ISOStringToSQLTimestamp(str)),
+  position: z.object({
+    lat: z
+      .string()
+      .transform((str) => Number(str))
+      .refine((num) => num >= -90, "should be greater than or equal to -90")
+      .refine((num) => num <= 90, "should be less than or equal to 90")
+      .or(z.number().gte(-90).lte(90)), // no need to transform if input is already a number (e.g. when coming from req. body)
+
+    long: z
+      .string()
+      .transform((str) => Number(str))
+      .refine((num) => num >= -180, "should be greater than or equal to -180")
+      .refine((num) => num <= 180, "should be less than or equal to 180")
+      .or(z.number().gte(-180).lte(180)), // no need to transform if input is already a number (e.g. when coming from req. body)
+  }),
+  sensors: z.array(
+    z.object({
+      id: z.number().positive(),
+      value: z.number(),
+      unit: z.string().optional(),
+    })
+  ),
+});
